// LibrarySystem_FrontV1.1/src/app/models/reservation.model.ts
export class ReservationModel{
    professionName: string;
    classesName: string;
    bookName: string;
    count: number;
    studentNo: string;
    isBorrowed: number;
    authorName: string;
    borrowDate: string;
    studentName: string;
    userId: string;
    remark: string;
    isbn: string;
    status: number;
    createTime: Date;
    isDelete: number;
    operator: string;
    updateTime: Date;
    id: string;
    location:string;
}